requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
Pygments==2.13.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
